arm (needs testing) and folder structure
