+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/tidev1_DO_NOT_USE/pidftuning/TuneElbow.java
@@ -27,56 +27,50 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.hardware;
+package org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.pidftuning;
 
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-public class Neck {
+import org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.Elbow;
 
-
-    // Define class members
-    double neckPosition = 0.5;
-    ElapsedTime neckDelay = new ElapsedTime();
-
-    private OpMode myOpMode;   // gain access to methods in the calling OpMode.
-
-    Servo servo = null;
-    public Neck(OpMode opmode) {
-        myOpMode = opmode;
-    }
-
-    public void init() {
-        // Define and Initialize Motors (note: need to use reference to actual OpMode).
-        servo = myOpMode.hardwareMap.get(Servo.class, "neck");
-
-    }
+/*
+ * This OpMode scans a single servo back and forward until Stop is pressed.
+ * The code is structured as a LinearOpMode
+ * INCREMENT sets how much to increase/decrease the servo position each cycle
+ * CYCLE_MS sets the update period.
+ *
+ * This code assumes a Servo configured with the name "left_hand" as is found on a Robot.
+ *
+ * NOTE: When any servo position is set, ALL attached servos are activated, so ensure that any other
+ * connected servos are able to move freely before running this test.
+ *
+ * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
+ * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
+ */
+@TeleOp(name = "Test: Elbow", group = "TunePIDF")
+public class TuneElbow extends LinearOpMode {
 
 
+    Elbow elbow = new Elbow(this);
 
-    public void listen() {
+    @Override
+    public void runOpMode() {
 
-        servo.setPosition(neckPosition);
-        // rotate right
-        if (myOpMode.gamepad2.right_bumper && neckDelay.seconds() > 1) {
-            neckDelay.reset();
-            neckPosition -=0.1;
-        // rotate left
-        } else if (myOpMode.gamepad2.left_bumper && neckDelay.seconds() > 1) {
-            neckDelay.reset();
-            neckPosition +=0.1;
+        // initialize all the hardware, using the hardware class. See how clean and simple this is?
+        elbow.init();
 
+
+        // Send telemetry message to signify robot waiting;
+        // Wait for the game to start (driver presses PLAY)
+        waitForStart();
+
+        // run until the end of the match (driver presses STOP)
+        while (opModeIsActive()) {
+
+            elbow.moveToDegree(90);
+            elbow.sendTelemetry();
+            updateTelemetry(telemetry);
         }
-    }
-
-
-    public void sendTelemetry() {
-        myOpMode.telemetry.addData("Neck Position", "%.2f", neckPosition);
     }
 }
