+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/tidev1_DO_NOT_USE/pidftuning/TuneViper.java
@@ -27,60 +27,50 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.hardware;
+package org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.pidftuning;
 
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-public class Mouth {
+import org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.Viper;
 
-
-    // Define class members
-
-
-
-    private OpMode myOpMode;   // gain access to methods in the calling OpMode.
-
-    double left_trigger;
-    double right_trigger;
-    double power;
-
-
-
-    CRServo mouth = null;
-    public Mouth(OpMode opmode) {
-        myOpMode = opmode;
-    }
-
-    public void init() {
-        // Define and Initialize Motors (note: need to use reference to actual OpMode).
-        mouth = myOpMode.hardwareMap.get(CRServo.class, "mouth");
+/*
+ * This OpMode scans a single servo back and forward until Stop is pressed.
+ * The code is structured as a LinearOpMode
+ * INCREMENT sets how much to increase/decrease the servo position each cycle
+ * CYCLE_MS sets the update period.
+ *
+ * This code assumes a Servo configured with the name "left_hand" as is found on a Robot.
+ *
+ * NOTE: When any servo position is set, ALL attached servos are activated, so ensure that any other
+ * connected servos are able to move freely before running this test.
+ *
+ * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
+ * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
+ */
+@TeleOp(name = "Test: Viper", group = "TunePIDF")
+public class TuneViper extends LinearOpMode {
 
 
-    }
+    Viper viper = new Viper(this);
+
+    @Override
+    public void runOpMode() {
+
+        // initialize all the hardware, using the hardware class. See how clean and simple this is?
+        viper.init();
 
 
+        // Send telemetry message to signify robot waiting;
+        // Wait for the game to start (driver presses PLAY)
+        waitForStart();
 
-    public void listen() {
-        left_trigger = myOpMode.gamepad2.left_trigger;
-        right_trigger = myOpMode.gamepad2.right_trigger;
+        // run until the end of the match (driver presses STOP)
+        while (opModeIsActive()) {
 
-        power = -left_trigger + right_trigger;
-
-        mouth.setPower(power);
-
-
-        myOpMode.telemetry.addData("Left Trigger", "%.2f", left_trigger);
-        myOpMode.telemetry.addData("Right Trigger", "%.2f", right_trigger);
-        myOpMode.telemetry.addData("Power", "%.2f", power);
-
-    }
-
-
-    public void sendTelemetry() {
-        myOpMode.telemetry.addData("Left Trigger", "%.2f", left_trigger);
-        myOpMode.telemetry.addData("Right Trigger", "%.2f", right_trigger);
-        myOpMode.telemetry.addData("Power", "%.2f", power);
+            viper.moveToDegree(90);
+            viper.sendTelemetry();
+            updateTelemetry(telemetry);
+        }
     }
 }
