+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/tidev1_DO_NOT_USE/tests/TestNeck.java
@@ -27,73 +27,35 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.hardware.tests;
+package org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import org.firstinspires.ftc.teamcode.hardware.tidev1_DO_NOT_USE.Neck;
 
-import org.firstinspires.ftc.teamcode.hardware.Elbow;
-import org.firstinspires.ftc.teamcode.hardware.Mouth;
-import org.firstinspires.ftc.teamcode.hardware.Neck;
-import org.firstinspires.ftc.teamcode.hardware.Viper;
-import org.firstinspires.ftc.teamcode.hardware.Wrist;
-
-/*
- * This OpMode scans a single servo back and forward until Stop is pressed.
- * The code is structured as a LinearOpMode
- * INCREMENT sets how much to increase/decrease the servo position each cycle
- * CYCLE_MS sets the update period.
- *
- * This code assumes a Servo configured with the name "left_hand" as is found on a Robot.
- *
- * NOTE: When any servo position is set, ALL attached servos are activated, so ensure that any other
- * connected servos are able to move freely before running this test.
- *
- * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
- * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
- */
-@TeleOp(name = "Test: All Attachments", group = "Concept")
-public class TestAttachments extends LinearOpMode {
+@TeleOp(name = "Test: Neck", group = "HardwareTest")
+public class TestNeck extends LinearOpMode {
 
 
-    Viper viper = new Viper(this);
-    Elbow elbow = new Elbow(this);
-    Mouth mouth = new Mouth(this);
     Neck neck = new Neck(this);
-    Wrist wrist = new Wrist(this);
 
     @Override
-    public void runOpMode() {
+    public void runOpMode() throws InterruptedException {
 
-        // initialize all the hardware, using the hardware class. See how clean and simple this is?
-        viper.init();
-        elbow.init();
-        mouth.init();
+        //starts hardware.
         neck.init();
-        wrist.init();
 
-        // Send telemetry message to signify robot waiting;
-        // Wait for the game to start (driver presses PLAY)
+
+
         waitForStart();
 
         // run until the end of the match (driver presses STOP)
         while (opModeIsActive()) {
 
-            viper.listen();
-            elbow.listen();
-            mouth.listen();
             neck.listen();
-            wrist.listen();
-
-
-            viper.sendTelemetry();
-            elbow.sendTelemetry();
-            mouth.sendTelemetry();
             neck.sendTelemetry();
-            wrist.sendTelemetry();
-
             updateTelemetry(telemetry);
-
         }
     }
 }
+
