+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/tidev2/Arm.java
@@ -27,20 +27,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.hardware;
+package org.firstinspires.ftc.teamcode.hardware.tidev2;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.CRServo;
 
-public class Elbow {
+public class Arm {
 
-    static final double ELBOW_UP_POS = -1;
-    static final double ELBOW_DOWN_POS = 1;
+    static final double ARM_UP_POS = -1;
+    static final double ARM_DOWN_POS = 1;
     static final int SLEEP_TIME = 500;
 
-    static final double COUNTS_PER_REVOLUTION = 5700.4; //Placeholder
-    static final double GEAR_RATIO = 50.9 / 1; //Placeholder
+    static final int COUNTS_PER_REVOLUTION = 288;
+    static final double GEAR_RATIO = 40 / 10;
 
 
     private double power_auto_move = 0.6;
@@ -53,13 +53,15 @@ public class Elbow {
     static final double  POWER_DOWN_MUL = 0.8;
     // Define class members
 
-    private DcMotorEx elbow;
 
     private OpMode myOpMode;   // gain access to methods in the calling OpMode.
 
     private double deg = 0.0;
 
-    public Elbow(OpMode opmode) {
+    private DcMotor arm_right = null;
+    private DcMotor arm_left = null;
+
+    public Arm (OpMode opmode) {
         myOpMode = opmode;
     }
 
@@ -69,29 +71,32 @@ public class Elbow {
 
     public void init() {
         // Define and Initialize Motors (note: need to use reference to actual OpMode).
-        elbow = myOpMode.hardwareMap.get(DcMotorEx.class, "elbow");
+        arm_right = myOpMode.hardwareMap.get(DcMotor.class, "arm_right");
+        arm_left = myOpMode.hardwareMap.get(DcMotor.class, "arm_left");
 
-        elbow.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        elbow.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        arm_right.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        arm_right.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
+        arm_left.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        arm_left.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
     }
 
-    public void moveElbowUp() {
+    public void moveArmUp() {
         deg = 150;
         moveToDegree(deg);
     }
 
-    public void moveElbowUpMore() {
+    public void moveArmUpMore() {
         deg = 180;
         moveToDegree(deg);
     }
 
-    public void moveElbowDown() {
+    public void moveArmDown() {
         deg = 0.0;
         moveToDegree(deg);
     }
 
-    public boolean isElbowUp() {
+    public boolean isArmUp() {
         if (deg >= 100) {
             return true;
         }
@@ -103,88 +108,100 @@ public class Elbow {
     }
 
     public double positionToDeg(int pos) {
-        return (double) elbow.getCurrentPosition() * 360  / (COUNTS_PER_REVOLUTION * GEAR_RATIO);
+        return (double) arm_right.getCurrentPosition() * 360  / (COUNTS_PER_REVOLUTION * GEAR_RATIO);
     }
 
     public void moveToDegree(double deg) {
         double targetPos = degToPosition(deg);
-//        boolean isGoingUp = targetPos > elbow.getCurrentPosition();
+//        boolean isGoingUp = targetPos > arm_right.getCurrentPosition();
         boolean isGoingUp = deg > 120;
 
 
-        elbow.setTargetPosition(((int)targetPos));
-        elbow.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        arm_right.setTargetPosition(((int)targetPos));
+        arm_right.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
 
+        arm_left.setTargetPosition(((int)targetPos));
+        arm_left.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
         // Set the required driving speed  (must be positive for RUN_TO_POSITION)
         // Start driving straight, and then enter the control loop
-        elbow.setPower(power_auto_move);
+        arm_right.setPower(power_auto_move);
+        arm_left.setPower(power_auto_move);
 
         // keep looping while we are still active, and BOTH motors are running.
-        while (elbow.isBusy()) {
+        while (arm_right.isBusy() && arm_left.isBusy()) {
 
             // decide if we reach a threshold to slow down
             if ((isGoingUp
-                    && (elbow.getCurrentPosition() > degToPosition(THRESHOLD_TO_SLOW_IN_DEG_HI)
-                    )
+                    && (arm_right.getCurrentPosition() > degToPosition(THRESHOLD_TO_SLOW_IN_DEG_HI)
+                    || arm_left.getCurrentPosition() > degToPosition(THRESHOLD_TO_SLOW_IN_DEG_HI))
             ) || ( !isGoingUp
-                    && (elbow.getCurrentPosition() < degToPosition(THRESHOLD_TO_SLOW_IN_DEG_LO)
-                    )
+                    && (arm_right.getCurrentPosition() < degToPosition(THRESHOLD_TO_SLOW_IN_DEG_LO)
+                    || arm_left.getCurrentPosition() < degToPosition(THRESHOLD_TO_SLOW_IN_DEG_LO))
             )) {
 
-                elbow.setPower(0.1);
+                arm_right.setPower(0.1);
+                arm_left.setPower(0.1);
             } else {
-                elbow.setPower(power_auto_move);
-                            }
+                arm_right.setPower(power_auto_move);
+                arm_left.setPower(power_auto_move);
+            }
 
             // Display drive status for the driver.
             sendTelemetry();
         }
 
         // Stop all motion & Turn off RUN_TO_POSITION
-        elbow.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        arm_right.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        arm_left.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        elbow.setPower(0);
+        arm_right.setPower(0);
+        arm_left.setPower(0);
+
         this.deg = deg;
 
     }
 
     public void sendTelemetry() {
-        myOpMode.telemetry.addData("Elbow Degree", "%4d",
-                elbow.getCurrentPosition());
+        myOpMode.telemetry.addData("Arm pos Left/Right", "%4d / %4d",
+                arm_left.getCurrentPosition(),
+                arm_right.getCurrentPosition());
     }
 
-    public void moveElbowByPower(double power) {
-        elbow.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+    public void moveArmByPower(double power) {
+        arm_right.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        arm_left.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
 
-
+        // TODO: allows pushing the arm down for now
         if (true || (power < 0) && (deg > 0)
                 || (power > 0) && (deg < 210)) {
-            elbow.setPower(power);
+            arm_right.setPower(power);
+            arm_left.setPower(power);
         }
 
-        deg = positionToDeg(elbow.getCurrentPosition());
+        deg = positionToDeg(arm_left.getCurrentPosition());
 
-        myOpMode.telemetry.addData("Elbow deg: ", "%.2f", deg);
+        myOpMode.telemetry.addData("Arm deg: ", "%.2f", deg);
         sendTelemetry();
     }
 
     public void listen() {
 
-        // move elbow according to the right stick y
+        // move arm according to the left stick y
 
 
         double power = -myOpMode.gamepad2.right_stick_y;
         if (Math.abs(power) > 0.1) {
-            moveElbowByPower(power);
+            moveArmByPower(power);
         } else {
-            elbow.setPower(0.0);
+            arm_right.setPower(0.0);
+            arm_left.setPower(0.0);
         }
 
         if (myOpMode.gamepad2.dpad_right) {
-            elbow.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
+            arm_right.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+            arm_left.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         }
 
     }
