+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/tidev2/tests/TestArm.java
@@ -27,38 +27,50 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.hardware.tests;
+package org.firstinspires.ftc.teamcode.hardware.tidev2.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.hardware.Elbow;
-import org.firstinspires.ftc.teamcode.hardware.Mouth;
-import org.firstinspires.ftc.teamcode.hardware.Neck;
+import org.firstinspires.ftc.teamcode.hardware.tidev2.Arm;
 
-@TeleOp(name = "Test: Neck", group = "HardwareTest")
-public class TestNeck extends LinearOpMode {
+/*
+ * This OpMode scans a single servo back and forward until Stop is pressed.
+ * The code is structured as a LinearOpMode
+ * INCREMENT sets how much to increase/decrease the servo position each cycle
+ * CYCLE_MS sets the update period.
+ *
+ * This code assumes a Servo configured with the name "left_hand" as is found on a Robot.
+ *
+ * NOTE: When any servo position is set, ALL attached servos are activated, so ensure that any other
+ * connected servos are able to move freely before running this test.
+ *
+ * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
+ * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
+ */
+@TeleOp(name = "Test: Arm", group = "Concept")
+public class TestArm extends LinearOpMode {
 
 
-    Neck neck = new Neck(this);
+    Arm arm = new Arm(this);
 
     @Override
-    public void runOpMode() throws InterruptedException {
+    public void runOpMode() {
 
-        //starts hardware.
-        neck.init();
+        // initialize all the hardware, using the hardware class. See how clean and simple this is?
+        arm.init();
 
 
-
+        // Send telemetry message to signify robot waiting;
+        // Wait for the game to start (driver presses PLAY)
         waitForStart();
 
         // run until the end of the match (driver presses STOP)
         while (opModeIsActive()) {
 
-            neck.listen();
-            neck.sendTelemetry();
-            updateTelemetry(telemetry);
+            arm.listen();
+            // Pace this loop so hands move at a reasonable speed.
+            sleep(50);
         }
     }
 }
-
